requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
python-dotenv==1.1.0
pymongo==4.6.1
email-validator==2.2.0